hide more button
