+++ src/utils/filterUtils.js
@@ -1,0 +1,11 @@
+export const filterMovies = (movies, searchQuery) => {
+  return movies.filter((movie) => {
+    return (
+      movie.nameRU.toLowerCase().indexOf(searchQuery) >= 0 || movie.nameEN?.toLowerCase().indexOf(searchQuery) >= 0
+    );
+  });
+};
+
+export const filterShortMovies = (movies, isSwitchChecked) => {
+  return movies.filter((movie) => (isSwitchChecked ? movie.duration <= 40 : Number));
+};
